Moved a lot of state management to CSS
Previously, there was a lot of state being managed for simple
interactions on the side. For example, determining if an accordion was
dropped open or closed. All of this is now being directly managed in
CSS. The advantage of this approach is that is cleans up the React code
and should be faster and more stable.

The styling of the website has also been changed.