feat: remove sidebar
This commit completely removes the sidebar component. It is replaced by
links to the "about me" and "blog" page in the header. There is now a
blog page that contains the list of blogs I've written. Previously, this
was included in a drop down menu in the sidebar. Removing the drop down
allowed me to delete a lot of JS logic and CSS configuration.